services/ids/requirements.txt
cryptography>=42.0.0
paho-mqtt>=2.0.0